update handle delete user and resto, update donnot delete default img when updating in menu and resto

+++ src/main/java/com/restoreserve/controlers/MenuController.java
@@ -128,7 +128,9 @@ public class MenuController {
                     Menu dataMenu = modelMapper.map(menuDto, Menu.class);
                     dataMenu.setRestaurant(dataRestaurant);
                     if (!menuDto.getPhoto().isEmpty()) {
-                        imageService.deleteImage(menuPrev.getPhoto());
+                        if (menuPrev.getPhoto() != "uploads/menu/default.png") {
+                            imageService.deleteImage(menuPrev.getPhoto());
+                        }
                         String imagePath = imageService.saveImage(menuDto.getPhoto(), "menu");
                         dataMenu.setPhoto(imagePath);
                     } else {
